ENH: move eigen_utils to common
