+++ common/so3_extra.h
