+++ include/Converter.h
@@ -9,7 +9,7 @@
 
 #include <Eigen/Dense>
 #include "sophus/se3.hpp"  //TODO: implement se3 by ourself to ensure the accuracy and rightness
-#include "so3_extra.h"
+#include "common/so3_extra.h"
 #ifdef USE_G2O_NEWEST
 #include "g2o/types/sba/types_six_dof_expmap.h"
 #include "g2o/types/sim3/types_seven_dof_expmap.h"
